+++ toolntip-core/toolntip-core.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Plugin Name: Toolntip Core
- * Plugin URI: https://toolntip.com
- * Description: Core functionality for the Toolntip platform.
- * Version: 1.0.60-dev.1
- * Author: Syed Asim Raza
- * Author URI: https://toolntip.com
+ * Plugin Name: ToolNTip Core
+ * Plugin URI: https://toolntip.com/
+ * Description: Core platform functionality for the ToolNTip ecosystem.
+ * Version: 1.0.60-dev.2
+ * Author: ToolNTip Technologies
+ * Author URI: https://toolntip.com/
  * License: GPL v2 or later
  * Text Domain: toolntip-core
  */
@@ -18,7 +18,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Plugin Constants
  */
 
-define( 'TNT_CORE_VERSION', '1.0.60-dev.1' );
+define( 'TNT_CORE_VERSION', '1.0.60-dev.2' );
 define( 'TNT_CORE_PATH', plugin_dir_path( __FILE__ ) );
 define( 'TNT_CORE_URL', plugin_dir_url( __FILE__ ) );